+++ macos/Classes/tools/NativeTools.m
@@ -13,33 +13,22 @@
 + (void)callPhone:(NSString *)phoneNumber {
     //    [[NSApplication sharedApplication] openURL:[NSURL URLWithString:[@"tel:://" stringByAppendingString:phoneNumber]]];
 }
-+ (NSMutableDictionary *)getAppInfo
-{
++ (NSDictionary *)getAppInfo{
     NSMutableDictionary *info = [NSMutableDictionary dictionary];
     NSDictionary *app = [[NSBundle mainBundle] infoDictionary];
-    //    CGRect statusBar = [[NSApplication sharedApplication] statusBarFrame];
-    //    [info setObject:@(statusBar.size.height) forKey:@"statusBarHeight"];
-    //    [info setObject:@(statusBar.size.width) forKey:@"statusBarWidth"];
-    
-    [info setObject:NSHomeDirectory() forKey:@"homeDirectory"];
-    [info setObject:[NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, YES) firstObject] forKey:@"documentDirectory"];
-    [info setObject:[NSSearchPathForDirectoriesInDomains(NSLibraryDirectory, NSUserDomainMask, YES) lastObject] forKey:@"libraryDirectory"];
-    [info setObject:[NSSearchPathForDirectoriesInDomains(NSCachesDirectory, NSUserDomainMask, YES) firstObject] forKey:@"cachesDirectory"];
-    [info setObject:NSTemporaryDirectory() forKey:@"temporaryDirectory"];
-    
-    [info setObject:[app objectForKey:@"CFBundleShortVersionString"] forKey:@"versionName"];
-    [info setObject:@"Apple" forKey:@"phoneBrand"];
-    [info setObject:[NSNumber numberWithInt:[[app objectForKey:@"CFBundleVersion"] intValue]] forKey:@"versionCode"];
-    
-    [info setObject:[app objectForKey:@"CFBundleIdentifier"] forKey:@"packageName"];
-    [info setObject:[app objectForKey:@"CFBundleName"] forKey:@"appName"];
-    [info setObject:[app objectForKey:@"DTSDKBuild"] forKey:@"sdkBuild"];
-    //    [info setObject:[app objectForKey:@"DTPlatformName"] forKey:@"platformName"];
-    //    [info setObject:[app objectForKey:@"MinimumOSVersion"] forKey:@"minimumOSVersion"];
-    [info setObject:[app objectForKey:@"DTPlatformVersion"] forKey:@"platformVersion"];
-    //    GDevice *device = [device currentDevice];
-    //    [info setObject:device->systemName forKey:@"systemName"];
-    //    [info setObject:device->systemVersion forKey:@"systemVersion"];
+    return(@{
+        @"homeDirectory" : NSHomeDirectory(),
+        @"documentDirectory" : [NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, YES) firstObject],
+        @"libraryDirectory" :[NSSearchPathForDirectoriesInDomains(NSLibraryDirectory, NSUserDomainMask, YES) lastObject],
+        @"cachesDirectory" : [NSSearchPathForDirectoriesInDomains(NSCachesDirectory, NSUserDomainMask, YES) firstObject],
+        @"temporaryDirectory" :NSTemporaryDirectory(),
+        @"versionName" : [app objectForKey:@"CFBundleShortVersionString"],
+        @"versionCode" : [NSNumber numberWithInt:[[app objectForKey:@"CFBundleVersion"] intValue]],
+        @"packageName" : [app objectForKey:@"CFBundleIdentifier"],
+        @"appName" : [app objectForKey:@"CFBundleName"],
+        @"sdkBuild" : [app objectForKey:@"DTSDKBuild"],
+        @"platformVersion" : [app objectForKey:@"DTPlatformVersion"],
+           });
     return info;
 }